+++ CHANGES.rst
@@ -1,7 +1,19 @@
 Changelog
 =========
 
-1.1 (2013-??-??)
+1.2 (2014-01-22)
+----------------
+
+* `Feature #17 <https://github.com/SolutionsCloud/apidoc/issues/17>`_ Merge ApiDoc's binaries
+* `Feature #18 <https://github.com/SolutionsCloud/apidoc/issues/18>`_ Merge cli arguments -f and -d in one argument -i (input)
+* `Feature #21 <https://github.com/SolutionsCloud/apidoc/issues/21>`_ Add patternProperties and additionalProperties in objects
+* `Feature #22 <https://github.com/SolutionsCloud/apidoc/issues/22>`_ "properties" and "items" are optionals
+* `Feature #23 <https://github.com/SolutionsCloud/apidoc/issues/23>`_ The property "type" is optional
+* `Bug #26 <https://github.com/SolutionsCloud/apidoc/issues/26>`_ Searchbox hide wrongly some elements
+* `Bug #27 <https://github.com/SolutionsCloud/apidoc/issues/27>`_ Keyboard shortcut for diff on search's filtered items
+* `Feature #32 <https://github.com/SolutionsCloud/apidoc/issues/32>`_ Enhances the theme of the readthedocs
+
+1.1 (2013-08-28)
 ----------------
 
 * `Feature #6 <https://github.com/SolutionsCloud/apidoc/issues/6>`_ Reduce resources size